Replace verif_lettres()
