--- verif_lettres.c
@@ -1,118 +1,0 @@
-#pragma once
-#include <stdio.h>
-#include <ctype.h>
-#include <string.h>
-
-#define MAX_JETON_TOUR 7
-#define MAX_JOUEUR 2
-
-char chaine_lettres_lues[MAX_JETON_TOUR];
-
-/**** Vérifie que les lettres saisies sont à la possession du joueur ****/
-int verif_lettres(int choix_joueur)
-{
-	int lettre = 'v';
-	int i;
-	int j = 0;
-	int trouve = 0;
-	int cpt = 0;						  // compteur pour vérifier que l'utilisateur n'a pas entré plus de 7 lettres
-	char tab_jeton_tempo[MAX_JETON_TOUR]; // Variable locale représentant le nombre de jetons du joueur.
-
-	// initialisation de la chaîne chaine_lettres_lues
-	for (i = 0; i < MAX_JETON_TOUR; i++)
-	{
-		chaine_lettres_lues[i] = '\0';
-	}
-
-	i = 0;
-
-	// saisie de chaine_lettres_lues
-
-	if ((lettre = getchar()) != '\n') // pour éviter des problèmes de buffer, si \n est toujours dans le buffer
-	{
-		if (lettre == '1')
-		{
-			return 2; // si lecture de "1", alors on quitte le menu d'échange
-		}
-		lettre = toupper(lettre);
-		chaine_lettres_lues[i] = lettre; // si \n n'est pas dans le buffer, alors il faut stocker la première lettre saisie directement avant de continuer
-		i++;
-		cpt++;
-		while ((lettre = getchar()) != '\n')
-		{
-			if (lettre != ' ') // si la lettre lue n'est pas un espace, on stocke la lettre dans chaine_lettre_lue
-			{
-				lettre = toupper(lettre);
-				chaine_lettres_lues[i] = lettre;
-				i++;
-				cpt++;
-			}
-		}
-	}
-
-	else // si le caractère '\n' n'est pas dans le buffer, on range les lettres saisies dans la variable chaine_lettre_lue
-	{
-		while ((lettre = getchar()) != '\n')
-		{
-			if (lettre == '1') // pour quitter le menu si 1 est saisi
-			{
-				return 2;
-			}
-			if (lettre != ' ')
-			{
-				lettre = toupper(lettre);
-				chaine_lettres_lues[i] = lettre;
-				i++;
-				cpt++;
-			}
-		}
-	}
-
-	// si cpt est supérieur au nombre max de jeton
-	if (cpt > MAX_JETON_TOUR)
-	{
-		printf("Erreur : vous avez entré un nombre de lettres supérieur aux jetons que vous possédez.\n");
-		chaine_lettres_lues[7] = '\0'; // on enlève la lettre résiduelle sinon elle reste
-		return trouve;
-	}
-
-	else
-	{
-
-		strcpy(tab_jeton_tempo, tabjoueur[choix_joueur].jeton); // pour comparer les lettres, utilisation d'une copie de la main du joueur
-		for (i = 0; chaine_lettres_lues[i] != '\0'; i++)		// parcours de chaine_lettre_lue jusqu'à '\0' qui correspond à une fin de saisie
-		{
-			if (chaine_lettres_lues[i] != '\0')
-			{
-				trouve = 0;
-				// à chaque lettre qu'on commence à vérifier, on remet trouve à 0
-				// pour vérifier que la lettre est bien dans la main du joueur
-				// si toutes les lettres sont bien en possession du joueur, on ne repassera pas dans ce if
-				// et trouve restera à 1
-			}
-			for (j = 0; j < MAX_JETON_TOUR; j++)
-			{
-				if (tab_jeton_tempo[j] == chaine_lettres_lues[i])
-				{
-					trouve = 1;
-					tab_jeton_tempo[j] = '\0'; // Suprression de la lettre du joueur dans tab_jeton_tempo pour éviter qu'elle soit relue par la suite
-					j = MAX_JETON_TOUR;		   // sortie de la boucle car la lettre est trouvée
-				}
-			}
-
-			if (trouve == 0) // Si une lettre n'a pas été trouvé, on sort de la boucle directement et la fonction renvoie trouve=0
-			{
-				printf("La lettre %c n'est pas dans votre jeu.\n", chaine_lettres_lues[i]);
-				return (trouve);
-			}
-		} // sortie boucle for
-
-		if (trouve == 1)
-		{
-			printf("Toutes les lettres entrées sont bien en votre possession.\n");
-		}
-
-	} // sortie du else
-
-	return trouve;
-}
